Optimizing code, correcting bugs and getting ready for translation

+++ src/consumer.cpp
@@ -16,17 +16,17 @@ Consumer::Consumer() : Worker() {
 
 bool Consumer::copyFile(QFile * source, QFile * destination, qint64 size) {
   char * bytes = new char[this->copyBufferSize];
-  qint64 actuallyWritten = 0, toBeWritten = 0, code = 0, totalWritten = 0;
+  qint64 actuallyWritten = 0, toBeWritten = 0, totalWritten = 0;
 
   if(source->open(QIODevice::ReadOnly))
-    this->log->logEvent("Opened source file: "+ source->fileName());
+    this->log->logEvent(tr("Successfully opened source file %1 for copying.").arg(source->fileName()));
   else
-    this->log->logError("Unable to open source file: " + source->fileName());
+    this->log->logError(tr("Failed to open source file %1 for copying.").arg(source->fileName()));
 
   if(destination->open(QIODevice::WriteOnly))
-    this->log->logEvent("Opened destination file: "+ destination->fileName());
+    this->log->logEvent(tr("Successfully opened destination file %1 for copying.").arg(destination->fileName()));
   else
-    this->log->logError("Unable to open destination file: " + destination->fileName());
+    this->log->logError(tr("Failed to open destination file %1 for copying.").arg(destination->fileName()));
 
   if(source->isReadable() && destination->isWritable()) {
     while((toBeWritten = source->read(bytes, this->copyBufferSize)) != 0) {
@@ -52,8 +52,10 @@ bool Consumer::copyFile(QFile * source, QFile * destination, qint64 size) {
     source->close();
     destination->close();
 
-    if(!(code = destination->setPermissions(source->permissions()))) {
-      this->log->logError("File copy failed (could not transfer file permissions)");
+    if(destination->setPermissions(source->permissions())) {
+      this->log->logEvent(tr("Permissions set successfully to the destination file %1.").arg(destination->fileName()));
+    } else {
+      this->log->logError(tr("Failed to set permissions to the destination file %1.").arg(destination->fileName()));
       return false;
     }
 
@@ -63,17 +65,20 @@ bool Consumer::copyFile(QFile * source, QFile * destination, qint64 size) {
     time.actime = s.lastRead().toSecsSinceEpoch();
     time.modtime = s.lastModified().toSecsSinceEpoch();
 
-    if((code = utime(destination->fileName().toStdString().c_str(), &time)) < 0){
-      this->log->logError("File copy failed (could not transfer file dates)");
+    if(utime(destination->fileName().toStdString().c_str(), &time) < 0){
+      this->log->logEvent(tr("Failed to transfer datetime to the destination file %1.").arg(destination->fileName()));
       return false;
+    } else {
+      this->log->logEvent(tr("Datetime transferred successfully to the destination file %1.").arg(destination->fileName()));
     }
 
+    this->log->logEvent(tr("Successfully copied file %1 to %2.").arg(destination->fileName()).arg(source->fileName()));
     return true;
   }
 
   error:
   delete[] bytes;
-  this->log->logError("File copy failed (error code: " + QString::number(code) + ")");
+  this->log->logError(tr("Failed to copy file %1 to %2.").arg(destination->fileName()).arg(source->fileName()));
 
   return false;
 }
@@ -129,7 +134,7 @@ void Consumer::work() {
   this->processedSize = 0;
   this->errorOccurred = false;
 
-  this->log->logEvent("Consumer has started");
+  this->log->logEvent(tr("Consumer process has begun working."));
 
   do {
     this->lock->lock();
@@ -160,18 +165,16 @@ void Consumer::work() {
       if(this->currentItem->getType() == TYPE_FILE) {
         if(this->currentFile->exists()) {
           if(this->currentItem->isSuperiorThan(existing, this->criterion)) {
-            emit this->triggerCurrentOperation(tr("Removing previous version of file"));
+            emit this->triggerCurrentOperation(tr("Removing older version of file"));
             if(this->currentFile->remove())
-              this->log->logEvent("Successfully removed previous version of file: " + existing);
+              this->log->logEvent(tr("Successfully removed older version of file %1.").arg(existing));
             else {
               this->errorOccurred = true;
-              this->log->logError("Failed to remove previous version of file: " + existing);
+              this->log->logError(tr("Failed to remove older version of file %1.").arg(existing));
             }
             goto copying;
           } else {
-            this->log->logEvent("Skipping item: " + this->currentItem->getName());
-            emit this->triggerCurrentOperation(tr("Skipping item"));
-            goto done;
+            goto skipping;
           }
         } else {
           goto copying;
@@ -179,9 +182,7 @@ void Consumer::work() {
       } else if(!this->currentDirectory->exists()) {
         goto cloning;
       } else {
-        this->log->logEvent("Skipping item: " + this->currentItem->getName());
-        emit this->triggerCurrentOperation(tr("Skipping item"));
-        goto done;
+        goto skipping;
       }
     } else {
       if(this->currentItem->getType() == TYPE_FILE) {
@@ -189,32 +190,33 @@ void Consumer::work() {
       } else if(this->currentItem->getType() == TYPE_DIRECTORY) {
         goto cloning;
       } else {
-        this->log->logEvent("Skipping item: " + this->currentItem->getName());
-        emit this->triggerCurrentOperation(tr("Skipping item"));
-        goto done;
+        goto skipping;
       }
     }
 
     copying:
     emit this->triggerCurrentOperation(tr("Copying file"));
 
-    if(this->copyFile(before, this->currentFile, this->currentItem->getSize()))
-      this->log->logEvent("Successfully copied file: " + left + " to " + existing);
-    else {
+    if(!this->copyFile(before, this->currentFile, this->currentItem->getSize()))
       this->errorOccurred = true;
-      this->log->logError("Failed to copy file: " + left + " to " + existing);
-    }
 
     goto done;
 
     cloning:
     emit this->triggerCurrentOperation(tr("Creating directory"));
-    if(beforeDirectory->mkdir(existing))
-      this->log->logEvent("Successfully cloned directory: " + left + " to " + existing);
-    else {
+
+    if(beforeDirectory->mkdir(existing)) {
+      this->log->logEvent(tr("Successfully recreated directory %1 as %2.").arg(left).arg(existing));
+    } else {
       this->errorOccurred = true;
-      this->log->logError("Failed to clone directory: " + left + " to " + existing);
+      this->log->logError(tr("Failed to recreate directory %1 as %2.").arg(left).arg(existing));
     }
+
+    goto done;
+
+    skipping:
+    this->log->logEvent(tr("Skipping item %1. The most recent version already exists in backup destination.").arg(this->currentItem->getName()));
+    emit this->triggerCurrentOperation(tr("Skipping item"));
 
     done:
     this->processedCount++;
@@ -245,26 +247,29 @@ void Consumer::work() {
 
     while(i.hasNext()) {
       QString current(i.next());
-      QString corresponding(this->source + "/" + t.relativeFilePath(current));
-      QFileInfo info(corresponding);
+      QFile corresponding(this->source + "/" + t.relativeFilePath(current));
+      QFileInfo info(current);
 
-      if(!info.exists()) {
+      if(!corresponding.exists() && info.exists()) {
         if(info.isDir()) {
           emit this->triggerCurrentOperation(tr("Removing folder"));
+
           QDir temp(current);
+
           if(temp.removeRecursively()) {
-            this->log->logEvent("Successfully removed obsolete folder: " + current);
+            this->log->logEvent(tr("Successfully removed obsolete folder %1.").arg(current));
           } else {
             this->errorOccurred = true;
-            this->log->logError("Failed to remove obsolete folder: " + current);
+            this->log->logError(tr("Failed to remove obsolete folder %1.").arg(current));
           }
         } else {
           emit this->triggerCurrentOperation(tr("Removing file"));
-          if(QFile::remove(current))
-            this->log->logEvent("Successfully removed obsolete file: " + current);
-          else {
+
+          if(QFile::remove(current)) {
+            this->log->logEvent(tr("Successfully removed obsolete file %1.").arg(current));
+          } else {
             this->errorOccurred = true;
-            this->log->logError("Failed to remove obsolete item: " + current);
+            this->log->logError(tr("Failed to remove obsolete file %1.").arg(current));
           }
         }
       }
@@ -273,5 +278,5 @@ void Consumer::work() {
 
   finish:
   emit this->finished();
-  this->log->logEvent("Consumer has finished");
+  this->log->logEvent(tr("Consumer process has finished working."));
 }
